pyproject.toml
[tool.poetry]
name = "fast-madr"
version = "0.1.0"
description = ""
authors = ["Mfreitas0 <mfreitasc0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
sqlalchemy = "^2.0.35"
taskipy = "^1.13.0"
pydantic = "^2.9.2"
psycopg = "^3.2.2"
pytest = "^8.3.3"
testcontainers = {extras = ["postgres"], version = "^4.8.1"}
alembic = "^1.13.3"
sqlacodegen = "3.0.0rc1"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.12"
pyjwt = "^2.9.0"
python-decouple = "^3.8"

[tool.taskipy.tasks]
run = 'fastapi dev fast_madr/main.py'
test = 'pytest -vv'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
